Laser wall and fixed moving platform
Laser wall now spawns as many segments as the script is told to, always spawning a start and an end segment. The segments and all the colliders will be periodically enabled and disabled according to the time variable.
Moving platform's vertical speed is now checked before applying on the player to avoid unwanted physics reactions