Basic door and treadmill prototype
Implemented a basic door with public methods to open, close, and toggle, with a basic animation state machine for the time being.
Refactored the MovingPlatform class to inherit from a new DynamicPlatform class to de couple the force movement logic and implement it in different classes like the treadmill. A new tag was also created to encompass all the children classes.
Prototype of the treadmill to implement and test forced horizontal movement when grounded. No sprites for the time being.