Dynamic complex collider system
Implemented a dynamic system for the player colliders that adjusts the colliders' size, offset and position according to the player's state. Multiple colliders are used to better accommodate the player's shape.
Overhauled the interactable base class to implement multiple types of interactions like interaction press or up button. The player's controller script has been adjusted to check for the type of interactable.